backend/pyproject.toml
[project]
name = "prelegal-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.1",
    "litellm>=1.82.1",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
